region-info: organize imports, update to scss, simplify link to interactive map, add comments
