+++ src/app/region-info/region-info.module.ts
@@ -1,11 +1,14 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-import { RegionInfoComponent } from './region-info/region-info.component';
-import { SharedModule } from '../shared/shared.module';
+import {
+  GeoserveCoreModule,
+  GeoserveOutputModule
+} from 'earthquake-geoserve-ui';
 
-import { GeoserveCoreModule, GeoserveOutputModule } from 'earthquake-geoserve-ui';
+import { SharedModule } from '../shared/shared.module';
+import { RegionInfoComponent } from './region-info/region-info.component';
 
 
 @NgModule({
