+++ src/app/region-info/region-info/region-info.component.scss
@@ -1,4 +1,6 @@
 shared-map {
+  cursor: pointer;
   display: block;
   height: 500px;
+  margin: 1em 0;
 }
