+++ src/app/region-info/region-info/region-info.component.spec.ts
@@ -1,14 +1,15 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { MockComponent } from 'ng2-mock-component';
 import { RouterTestingModule } from '@angular/router/testing';
 
-import { RegionInfoComponent } from './region-info.component';
-import { Event } from '../../event';
-import { of } from 'rxjs/observable/of';
-import { EventService } from '../../../..';
-import { MockPipe } from '../../mock-pipe';
-
 import { CoordinatesService } from 'earthquake-geoserve-ui';
+import { MockComponent } from 'ng2-mock-component';
+import { of } from 'rxjs/observable/of';
+
+import { EventService } from '../../core/event.service';
+import { Event } from '../../event';
+import { MockPipe } from '../../mock-pipe';
+import { RegionInfoComponent } from './region-info.component';
+
 
 describe('RegionInfoComponent', () => {
   let component: RegionInfoComponent;
