+++ src/app/region-info/region-info/region-info.component.ts
@@ -1,30 +1,32 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 
+import { CoordinatesService } from 'earthquake-geoserve-ui';
+import * as L from 'leaflet';
 import { Subscription } from 'rxjs';
 
-import { HistoricSeismicityOverlay } from '../../shared/map-overlay/historic-seismicity-overlay';
-import { EventService } from '../../../..';
-
-import { CoordinatesService } from 'earthquake-geoserve-ui';
-
-import * as L from 'leaflet';
+import { EventService } from '../../core/event.service';
 
 
+/**
+ * Displays regional information related to the event epicenter and displays
+ * a static map. The static map displays the event epicenter and historic
+ * seismicity overlays while linking to the interactive map with the same
+ * overlays enabled.
+ */
 @Component({
   selector: 'app-region-info',
   templateUrl: './region-info.component.html',
-  styleUrls: ['./region-info.component.css']
+  styleUrls: ['./region-info.component.scss']
 })
 export class RegionInfoComponent implements OnDestroy, OnInit {
+  public subscription: Subscription;
 
-  public overlays: Array<L.Layer> = [ new HistoricSeismicityOverlay() ];
-
-  private subscription: Subscription;
 
   constructor (
     public coordinatesService: CoordinatesService,
     public eventService: EventService
   ) { }
+
 
   ngOnInit () {
     this.subscription = this.eventService.event$.subscribe((event: any) => {
@@ -36,6 +38,14 @@ export class RegionInfoComponent implements OnDestroy, OnInit {
     this.subscription.unsubscribe();
   }
 
+  /**
+   * Updates the coordinate service with the event epicenter. The coordinate
+   * service then triggers a fetch of the geoserve regional information, for:
+   * administrative regions, nearby places, and tectonic summary.
+   *
+   * @param event {any}
+   *    An event object used to define the epicenter
+   */
   updateGeoserveCoordinateService (event: any) {
     if (!event || !event.geometry) {
       return;
